requirements.txt
Django==5.1.4
django-ninja==1.3.0
python-decouple==3.8
qrcode==8.0
pillow==11.1.0
